requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.26.5

# CLI Tools
typer==0.12.3
httpx==0.25.2
rich==13.7.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
ruff==0.1.7
mypy==1.7.1
isort==5.12.0

# Development
python-dotenv==1.0.0
django-extensions==3.2.3 